requirements.txt
annotated-types==0.7.0
anyio==3.7.1
beautifulsoup4==4.12.2
bs4==0.0.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
fastapi==0.104.1
feedparser==6.0.10
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
psycopg2-binary==2.9.9
pydantic==2.11.10
pydantic-settings==2.11.0
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.6
PyYAML==6.0.3
requests==2.31.0
sgmllib3k==1.0.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.23
starlette==0.27.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
